+++ eras/babbage/impl/test/Main.hs
@@ -6,7 +6,6 @@ module Main where
 import Cardano.Ledger.Babbage (BabbageEra)
 import qualified Test.Cardano.Ledger.Alonzo.Binary.CostModelsSpec as CostModelsSpec
 import qualified Test.Cardano.Ledger.Alonzo.Binary.TxWitsSpec as TxWitsSpec
-import qualified Test.Cardano.Ledger.Alonzo.Imp as Alonzo
 import qualified Test.Cardano.Ledger.Babbage.Binary.CddlSpec as CddlSpec
 import qualified Test.Cardano.Ledger.Babbage.BinarySpec as BinarySpec
 import qualified Test.Cardano.Ledger.Babbage.GoldenSpec as GoldenSpec
@@ -17,14 +16,13 @@ import qualified Test.Cardano.Ledger.Babbage.TxInfoSpec as TxInfo
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import Test.Cardano.Ledger.Era
-import qualified Test.Cardano.Ledger.Shelley.Imp as Shelley
 import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 instance EraSpec BabbageEra where
   eraImpSpec era = do
-    Shelley.shelleyEraSpecificSpec era
-    Alonzo.alonzoEraSpecificSpec era
-    Imp.babbageEraSpecificSpec era
+    Imp.shelleyToBabbageSpec era
+    Imp.alonzoToConwaySpec era
+    Imp.babbageOnlySpec era
     Imp.spec era
 
 main :: IO ()