+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp.hs
@@ -7,7 +7,11 @@
 {-# LANGUAGE TypeOperators #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Conway.Imp (spec, conwayEraSpecificSpec) where
+module Test.Cardano.Ledger.Conway.Imp (
+  spec,
+  Babbage.alonzoToConwaySpec,
+  conwayOnlySpec,
+) where
 
 import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Core
@@ -61,9 +65,9 @@ spec era = do
     UTXOS.spec
     UTXOW.spec
 
-conwayEraSpecificSpec :: Spec
-conwayEraSpecificSpec = do
+conwayOnlySpec :: Spec
+conwayOnlySpec = do
   describe "ConwayEra Specific" $ withImpInitEachEraVersion (Proxy @ConwayEra) $ do
     -- TODO: move to `spec` when ready: https://github.com/IntersectMBO/cardano-ledger/issues/5805
     CERTS.spec
-    UTXO.conwayEraSpecificSpec
+    UTXO.conwayOnlySpec