+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp/UtxoSpec.hs
@@ -10,7 +10,7 @@
 
 module Test.Cardano.Ledger.Conway.Imp.UtxoSpec (
   spec,
-  conwayEraSpecificSpec,
+  conwayOnlySpec,
 ) where
 
 import Cardano.Ledger.Address
@@ -151,13 +151,13 @@ spec = describe "UTXO" $ do
           ++ extraScripts
           ++ extraScripts
 
-conwayEraSpecificSpec ::
+conwayOnlySpec ::
   forall era.
   ( ConwayEraImp era
   , ShelleyEraTxCert era
   ) =>
   SpecWith (ImpInit (LedgerSpec era))
-conwayEraSpecificSpec = describe "UTXO" $ do
+conwayOnlySpec = describe "UTXO" $ do
   describe "Certificates" $ do
     it "Reg/UnReg collect and refund correct amounts" $ do
       utxoStart <- getUTxO